+++ fj-doc-val-pdfbox/src/main/java/org/fugerit/java/doc/val/pdf/box/PdfboxStrictValidator.java
@@ -25,6 +25,7 @@ public class PdfboxStrictValidator extends AbstractDocTypeValidator {
 			result = DocTypeValidationResult.newOk();
 		} catch (Exception e) {
 			log.warn( "Failed check on pdf : {}", e.toString() );
+			result.withMainException( e );
 		}
 		return result;
 	}