+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- [fj-doc-val-*] DocTypeValidationResult has now validation message and exceptions
+
 ### Changed
 
 - quarkus-version set to 3.24.3 across all the modules <https://github.com/fugerit-org/fj-doc/issues/469>